Moved shortener code outside of shortify transformer
