+++ lib/index.js
@@ -1,29 +1,10 @@
 'use strict';
 
 var through = require('through');
-
-var PAIRS = [
-  // `require(""`
-  {
-    query: ['require[(]([\'"])([^\'"]*)', '([^\'"]*)([\'"])'],
-    replace: ['require($1$2', '$3$4']
-  },
-
-  // ` from ""`
-  {
-    query: ['[\\s]from[\\s]([\'"])([^\'"]*)', '([^\'"]*)([\'"])'],
-    replace: [' from $1$2', '$3$4']
-  },
-
-  // `import "";`
-  {
-    query: ['import[\\s]([\'"])([^\'"]*)', '([^\'"]*)([\'"])'],
-    replace: ['import $1$2', '$3$4']
-  }
-];
+var shortenerBuilder = require('./shortener.js');
 
 module.exports = function (options) {
-  options = options || {};
+  var shortener = shortenerBuilder(options || {});
 
   return function (filename) {
     var source = '';
@@ -33,21 +14,7 @@ module.exports = function (options) {
     }
 
     function end() {
-      for (var key in options) {
-        if (!options.hasOwnProperty(key)) {
-          continue;
-        }
-        PAIRS.forEach(function (params) {
-          var query = params.query;
-          var replace = params.replace;
-          var searchTerm = query[0] + key + query[1];
-          var replaceTerm = replace[0] + options[key] + replace[1];
-          var regex = new RegExp(searchTerm, 'g');
-          source = source.replace(regex, replaceTerm);
-        });
-      }
-
-      this.queue(source);
+      this.queue(shortener(source));
       this.queue(null);
     }
